Rename icall-def-core.h to icall-def-netcore.h.
