+++ mono/metadata/icall-def-netcore.h
